% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_as_list.R
\name{download_as_list}
\alias{download_as_list}
\title{Download results from ocpu server as a list}
\usage{
download_as_list(resp, pattern = "R/.val")
}
\arguments{
\item{resp}{An object of class
\code{\link[httr:response]{response}}, typically returned by a
previous call to \code{httr::POST} to the server.}

\item{pattern}{A character vector with the trailing part of the requested
url, e.g. \code{"R/.val"} or \code{"stdout"}. Use \code{""} to
obtain all results.}
}
\value{
A list with the contents under the url's that match
  \code{pattern}. The result has length 0 if the pattern was
  not found.
}
\description{
Obtains thecontents of one or more url's from the OpenCPU server response
in a list.
}
\examples{
library(jamesclient)
file <- file.path(path.package("jamesclient"), "testdata", "client3.json")
resp <- upload_bds(file)
resp
download_as_list(resp, pattern = "stdout")
}
