% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_location}
\alias{extract_location}
\alias{extract_session}
\alias{extract_svg}
\alias{extract_return}
\title{Extract useful parts from OpenCPU response}
\usage{
extract_location(resp)

extract_session(resp)

extract_svg(resp, width = 7, height = 7)

extract_return(resp)
}
\arguments{
\item{resp}{An object of class \code{\link[httr:response]{response}}}

\item{width}{the width of the device in inches.}

\item{height}{the height of the device in inches.}
}
\value{
A url
}
\description{
\code{extract_location()} extracts the location
}
