% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_return_url.R
\name{extract_return_url}
\alias{extract_return_url}
\title{Extract url of function return on server}
\usage{
extract_return_url(resp)
}
\arguments{
\item{resp}{An object of class \code{\link[httr:response]{response}}}
}
\value{
A url ending in \code{"R/.val"}
}
\description{
Extract url of function return on server
}
\keyword{client}
