% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_url.R
\name{get_url}
\alias{get_url}
\alias{get_url_svg}
\title{Get url from OpenCPU response}
\usage{
get_url(resp, name = c("return", "location", "session", "console",
  "stdout", "svg", "messages"), ...)

get_url_svg(resp, pad = "?width=7&height=7", ...)
}
\arguments{
\item{resp}{An object of class \code{\link[httr:response]{response}}
returned by OpenCPU.}

\item{name}{A string: \code{"location"}, \code{"session"},
\code{"stdout"}, \code{"svg"} or \code{"return"}. The default is
\code{"return"}.}

\item{\dots}{Additional string that is concatenate to the URL}

\item{pad}{A string to be padded to the url.}
}
\value{
A url. If not found, the return is \code{character(0)}.
}
\description{
Get url from OpenCPU response
}
\details{
Only \code{get_url()} is exported, so use \code{get_url()}
in your code.
}
