% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_upload.R
\name{inspect_demodata}
\alias{inspect_demodata}
\title{Inspect demo data uploaded to JAMES}
\usage{
inspect_demodata(
  name,
  cabinet,
  format = c("2.0", "1.0", "1.1"),
  host = "https://groeidiagrammen.nl"
)
}
\arguments{
\item{name}{Name of the child}

\item{cabinet}{Name of cabinet}

\item{format}{String. JSON data schema version number. There are currently
three schemas supported: \code{1.0}, \code{1.1} and \code{2.0}.}

\item{host}{Full host name}
}
\description{
Uploads JSON child demo data to JAMES and downloads the processed data for inspection.
}
\examples{
data <- inspect_demodata(name = "Anne_S", "smocc", format = "2.0", host = "http://localhost")
head(data)
}
