% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/james_get.R
\name{james_get}
\alias{james_get}
\title{Make JAMES GET request}
\usage{
james_get(host = "http://localhost", path = character(0), ...)
}
\arguments{
\item{host}{String with the host. The default is \code{"http://localhost"}.}

\item{path}{String with the path to the called end point,
e.g. \code{"version"}, \code{"data/upload"} or \code{"data/upload/json"}.}

\item{\dots}{Extra arguments to create the URI in GET()}
}
\value{
Object of class \code{james_get}
}
\description{
Make JAMES GET request
}
