% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_json.R
\name{read_json_js}
\alias{read_json_js}
\alias{read_json_jo}
\title{Read JSON content from file or URL}
\usage{
read_json_js(txt = NULL)

read_json_jo(txt = NULL)
}
\arguments{
\item{txt}{A file name or URL}
}
\description{
Both \code{read_json_js()} and \code{read_json_jo()} read JSON from a file or an URL.
Function \code{read_json_js()} returns a string, while \code{read_json_jo()} returns
an object of class \code{jsonlite::class-json}.
}
\examples{
url <- paste("http://localhost", "ocpu/library/jamesdemodata",
  "extdata/bds_v3.0/smocc/Laura_S.json", sep = "/")
fn <- system.file("extdata", "bds_v3.0", "smocc", "Laura_S.json",
 package = "jamesdemodata", mustWork = TRUE)

js1 <- read_json_js(url)
js2 <- read_json_js(fn)
jo1 <- read_json_jo(url)
jo2 <- read_json_jo(fn)
}
