% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req_chart.R
\name{req_chart}
\alias{req_chart}
\title{Request chart from server}
\usage{
req_chart(location, chartcode = NULL, curve_interpolation = TRUE)
}
\arguments{
\item{location}{An url pointing the server-side data from
upload.}

\item{chartcode}{The code of the requested growth chart. If not
specified, the server will automatically plot child height for
the most recent age period.}

\item{curve_interpolation}{A logical indicating whether curve interpolation shoud be applied.}
}
\value{
An object of class \code{\link[httr:response]{response}}
}
\description{
This function assumes that the data are stored on the server.
}
\keyword{client}
