% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_chart.R
\name{request_chart}
\alias{request_chart}
\title{Request a growth chart from groeidiagrammen.nl}
\usage{
request_chart(location, chartcode = NULL, curve_interpolation = TRUE)
}
\arguments{
\item{location}{An url pointing the server-side data of
class `individual`.}

\item{chartcode}{The code of the requested growth chart. If not
specified, the server will automatically plot child height for
the most recent age period.}

\item{curve_interpolation}{A logical indicating whether curve interpolation shoud be applied.}
}
\value{
An object of class \code{\link[httr:response]{response}}
}
\description{
This function requests a growth chart from \code{groeidiagrammen.nl}.
The function assumes that the data are stored on the server.
}
\examples{
library("james.client")
file <- file.path(path.package("james.client"), "testdata", "client3.json")
resp <- upload_bds(file)
loc <- extract_location(resp)
resp2 <- request_chart(loc, "NJAA")
resp2
}
\seealso{
\code{\link[james]{draw_chart_ind}}, \code{\link[httr]{POST}}
}
\keyword{client}
