% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_site.R
\name{request_site}
\alias{request_site}
\title{Request a site personalised chart and user nagivation}
\usage{
request_site(bds = NULL, ssd = NULL,
  host = "http://groeidiagrammen.nl", path = "ocpu/library/james")
}
\arguments{
\item{bds}{Name of file that contains valid JSON data. The
variable specification are expected to be according specification
\href{https://www.ncj.nl/themadossiers/informatisering/basisdataset/documentatie/?cat=12}{BDS
 JGZ 3.2.5}, and converted to JSON.}

\item{ssd}{Server-side data location, for example, as obtained
by \code{get_url(resp, "location")}, where \code{resp} is the
object returned by OpenCPU server.}

\item{host}{A url of the JAMES server host.}

\item{path}{Path to application on JAMES server.}
}
\value{
An object of class \code{\link[httr:response]{response}}
}
\description{
This function requests an URL to a site that shows a personalised
growth chart plus essential user nagivation from JANES. The
function can take child data in two forms: 1) a JSON file with
BDS-formatted child data or, 2) a URL with child data on a
previously stored server location on the server.
}
\details{
One of \code{txt} or \code{resp} need to be specified. If both
are given, a non-NULL \code{txt} takes precedence over \code{resp}.
}
\examples{
library("jamesclient")
fn <- file.path(path.package("jamesclient"), "testdata", "client3.json")

# as two steps
resp <- upload_bds(fn)
ssd <- get_url(resp, "location")
site_url <- request_site(ssd = ssd)
site_url
# browseURL(site_url)

# into one step
site_url <- request_site(bds = fn)
site_url

\dontrun{
# using localhost
# request_site does not work on localhost, presumably
# because the local machine does not run a full webserver
# however, if file is uploaded to groeidiagrammen.nl,
# then the localhost can download data from there

# --- this does not work
r1 <- upload_bds(fn,
                 host = "http://localhost:5656",
                 path = "ocpu/apps/stefvanbuuren/james")
ssd <- get_url(r1, "location")
site_url <- request_site(ssd = ssd,
                         host = "http://localhost:5656",
                         path = "ocpu/apps/stefvanbuuren/james")
site_url
# browseURL(site_url)

# --- this will work
r1 <- upload_bds(fn)
ssd <- get_url(r1, "location")
site_url <- request_site(ssd = ssd,
                         host = "http://localhost:5656",
                         path = "ocpu/apps/stefvanbuuren/james")
site_url
# browseURL(site_url)
}
}
\seealso{
\code{\link[james]{draw_chart}},
\code{\link{upload_bds}}.
}
\keyword{client}
