% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_chart.R
\name{update_chart}
\alias{update_chart}
\title{Update chart using data stored on server}
\usage{
update_chart(location, chartcode = NULL, curve_interpolation = TRUE)
}
\arguments{
\item{location}{An url pointing the server-side data from
data upload. The result of a \code{POST} call.}

\item{chartcode}{The code of the requested growth chart. If not
specified, the server will automatically plot child height for
the most recent age period.}

\item{curve_interpolation}{A logical indicating whether curve interpolation shoud be applied.}
}
\value{
An object of class \code{\link[httr:response]{response}}
}
\description{
This function assumes that the data are stored on the server.
}
\seealso{
\code{\link[james]{draw_chart_ind}}, \code{\link[httr]{POST}}
}
\keyword{client}
