% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_bds.R
\name{upload_bds}
\alias{upload_bds}
\title{Upload and parse BDS file to JAMES}
\usage{
upload_bds(bds, host = "https://groeidiagrammen.nl",
  path = "ocpu/library/james/R/convert_bds_ind")
}
\arguments{
\item{bds}{File name or a string referring to the data (in JSON
format) to be uploaded. The variable specification are expected
to be according specification
\href{https://www.ncj.nl/themadossiers/informatisering/basisdataset/documentatie/?cat=12}{BDS
 JGZ 3.2.5}, and converted to JSON.}

\item{host}{String with URL of the JAMES host machine. Defaults to
\code{https://groeidiagrammen.nl}.}

\item{path}{Path to JAMES application on host. Defaults to
\code{ocpu/library/james}.}
}
\value{
An object of class \code{\link[httr:response]{response}}
}
\description{
Uploads a JSON file or string with BDS-coded data, parses its
contents into an object of class \code{individual}. The
function returns various url's on the uploaded data.
}
\note{
At present there is only minimal error checking on the BDS file.
}
\examples{
fn <- system.file("extdata", "allegrosultum", "client3.json", package = "jamestest")

# upload JSON file to groeidiagrammen.nl
r1 <- upload_bds(fn)
httr::status_code(r1)

# upload JSON file to localhost
host <- "http://localhost:5656"
path <- "ocpu/apps/stefvanbuuren/james/R/convert_bds_ind"
r2 <- upload_bds(fn, host = host, path = path)
httr::status_code(r2)

# upload JSON string
js <- jsonlite::toJSON(jsonlite::fromJSON(fn), auto_unbox = TRUE)
r3 <- upload_bds(js)
httr::status_code(r3)

# upload JSON from URL
url <- "https://groeidiagrammen.nl/ocpu/library/james/testdata/client3.json"
r4 <- upload_bds(url)
httr::status_code(r4)

\dontrun{
# the following does not yet work
r5 <- upload_bds(url, host = host, path = path)
}
}
\seealso{
\code{\link[minihealth]{convert_bds_individual}},
\code{\link{request_site}}, \code{\link[httr:response]{response}}
}
\keyword{client}
