% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_txt.R
\name{upload_txt}
\alias{upload_txt}
\title{Upload BDS JSON Data to JAMES Server}
\usage{
upload_txt(
  txt,
  host = "http://localhost:8080",
  format = "3.0",
  schema = NULL,
  verbose = FALSE
)
}
\arguments{
\item{txt}{The input data to upload. Can be a local JSON file path, a URL
pointing to JSON data, or a JSON string or object.}

\item{host}{Character. URL of the JAMES server. Default: \code{"http://localhost:8080"}.}

\item{format}{Character. JSON schema version number. Supported formats are
\code{"1.0"}, \code{"1.1"}, \code{"2.0"}, and \code{"3.0"}. Format \code{"3.0"} is recommended.
If \code{schema} is provided, it overrides \code{format}.}

\item{schema}{Optional character. File path to a JSON schema for validation.
Overrides \code{format}. The version number is extracted from the file name.}

\item{verbose}{Logical. If \code{TRUE}, prints diagnostic output of the POST request.}
}
\value{
An object of class \code{\link[httr:response]{httr::response()}} containing the server's response.
}
\description{
Uploads and validates BDS data from a JSON file, URL, or string, sends it to a
JAMES server, and returns the server response. On success, the data is stored
as a tibble with a person attribute on the host.
}
\details{
If an object named \code{ua} (a user agent created via \code{\link[httr:user_agent]{httr::user_agent()}})
exists in the environment, it will be added to the request headers.
While optional, identifying yourself can be useful for diagnostics.

To retrieve a partial JSON representation of the internal S4 class \code{individual},
append \code{"/json"} to the \code{path} and set \code{query = list(auto_unbox = TRUE, force = TRUE)}.
Note: This representation is not intended for reconstructing the class, and may
change without notice. Use dedicated APIs to access internal data elements such
as Z-scores or brokenstick estimates.
}
\examples{
host <- "https://james.groeidiagrammen.nl"
url <- paste0("https://raw.githubusercontent.com/growthcharts/jamesdemodata/",
              "refs/heads/master/inst/json/examples/Laura_S.json")
fn <- system.file("extdata", "bds_v3.0", "smocc", "Laura_S.json",
                  package = "jamesdemodata", mustWork = TRUE)
js <- read_json_js(url)
jo <- read_json_jo(url)

r1 <- upload_txt(fn, host)
stopifnot(httr::status_code(r1) == 201L)

r2a <- upload_txt(js, host)
r2b <- upload_txt(jo, host)

r3 <- upload_txt(url, host, format = "1.0")
stopifnot(httr::status_code(r3) == 201L)

# View server response URLs
# browseURL(file.path(get_url(r3), "print"))
# browseURL(file.path(get_url(r3), "json"))

}
