% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_url.R
\name{valid_url}
\alias{valid_url}
\title{Check for URL existence}
\usage{
valid_url(url_in, t = 2)
}
\arguments{
\item{url_in}{String of URL to test}

\item{t}{Time out}
}
\description{
Taken from \url{https://stackoverflow.com/questions/52911812/check-if-url-exists-in-r}.
}
\examples{
urls <- c(
  "http://www.amazon.com", "http://this.isafakelink.biz",
  "https://stackoverflow.com", "http://localhost:8080")
sapply(urls, valid_url)
}
