% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_key.R
\name{extract_location}
\alias{extract_location}
\alias{extract_svg}
\title{Extract location of the response on the server}
\usage{
extract_location(resp)

extract_svg(resp, width = 7, height = 7)
}
\arguments{
\item{resp}{An object of class \code{\link[httr:response]{response}}}

\item{width}{the width of the device in inches.}

\item{height}{the height of the device in inches.}
}
\value{
A url

A url
}
\description{
Extract location of the response on the server

Extract svg url on OpenCPU server
}
\keyword{client}
