% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_result.R
\name{get_result}
\alias{get_result}
\title{Get result from ocpu server}
\usage{
get_result(resp, pattern = "R/.val")
}
\arguments{
\item{resp}{An object of class
\code{\link[httr:response]{response}}, typically returned by a
previous call to \code{httr::POST} to the server.}

\item{pattern}{A string with the trailing part of the requested
url, e.g. \code{"R/.val"} or \code{"stdout"}. Use \code{""} to
obtain all results.}
}
\value{
A list with the contents under the url's that match
  \code{pattern}. The result has length 0 if the pattern was
  not found.
}
\description{
Obtains the contents of one or more url's from a server response
in a two-step communication.
}
\examples{
library(james.client)
file <- file.path(path.package("james.client"), "testdata", "client3.json")
resp <- upload_bds(file)
resp
get_result(resp, pattern = "stdout")
}
