% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_demodata.R
\name{inspect_demodata}
\alias{inspect_demodata}
\title{Inspect demo data uploaded to JAMES}
\usage{
inspect_demodata(
  name,
  cabinet,
  format = c("3.0", "2.0", "1.0", "1.1"),
  host = "http://localhost"
)
}
\arguments{
\item{name}{Name of the child}

\item{cabinet}{Name of cabinet}

\item{format}{String. JSON data schema version number. There are currently
three schemas supported: \code{"1.0"}, \code{"1.1"} and \code{"2.0"}. Formats \code{"1.0"} and
\code{"1.1"} are included for backward compatibility only. Use \code{format = "2.0"}
for new applications.}

\item{host}{String with the host. The default is \code{"http://localhost"}.}
}
\description{
Uploads JSON child demo data to localhost and downloads the processed data
for inspection.
}
\examples{
\dontrun{
data <- inspect_demodata(name = "Anne_S", cabinet = "smocc")
head(data)
}
}
