% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_upload.R
\name{inspect_upload}
\alias{inspect_upload}
\title{Inspect child data uploaded to JAMES}
\usage{
inspect_upload(
  name,
  cabinet,
  version = 2L,
  host = "https://groeidiagrammen.nl"
)
}
\arguments{
\item{name}{Name of the child}

\item{cabinet}{Name of cabinet}

\item{version}{Integer. JSON schema version number. There are currently two
schemas supported, versions \code{1} and \code{2}.}

\item{host}{Full host name}
}
\description{
Uploads JSON child data to JAMES and downloads the processed data for inspection.
}
\examples{
data <- inspect_upload(name = "Anne_S", "smocc", version = 1)
head(data)
}
