% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_upload.R
\name{inspect_upload}
\alias{inspect_upload}
\title{Inspect child data uploaded to JAMES}
\usage{
inspect_upload(
  name,
  cabinet,
  schema = "bds_str",
  host = "https://groeidiagrammen.nl"
)
}
\arguments{
\item{name}{Name of the child}

\item{cabinet}{Name of cabinet}

\item{schema}{Name of schema, default \code{bds_str}}

\item{host}{Full host name}
}
\description{
Uploads JSON child data to JAMES and downloads the processed data for inspection.
}
\examples{
data <- inspect_upload(name = "Anne_S", "smocc")
data
}
