% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/james_api.R
\name{james_get}
\alias{james_get}
\title{Make JAMES GET request}
\usage{
james_get(host = "http://localhost", path = character(0))
}
\arguments{
\item{host}{String with the host. The default is \code{"http://localhost"}}

\item{path}{String with the path}
}
\value{
Object of class \code{james_get}
}
\description{
Make JAMES GET request
}
