% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/james_post.R
\name{james_post}
\alias{james_post}
\title{Make JAMES POST request}
\usage{
james_post(host = "http://localhost", path = character(0), txt = NULL, ...)
}
\arguments{
\item{host}{String with the host. The default is \code{"http://localhost"}}

\item{path}{String with the path, e.g. \code{"version"}, \code{"upload"} or \code{"upload/json"}}

\item{txt}{Data set, used in \code{"upload"}}

\item{\dots}{Any other arguments passed to james functions via POST body}
}
\value{
Object of class \code{james_post}
}
\description{
Make JAMES POST request
}
\details{
If \code{txt} is a file, then the data are uploaded using \code{upload_file()} and
encoding \code{multipart/form}. If \code{txt} is a URL, then the data are read from the URL
as JSON string and uploaded with encoding \code{json}.
}
\examples{
fn <- system.file("extdata", "allegrosultum", "client3.json", package = "jamesdemodata")
url <- paste("http://localhost", "ocpu/library/jamesdemodata",
  "extdata/bds_v2.0/smocc/Laura_S.json", sep = "/")
fn1 <- system.file("extdata", "bds_v2.0", "smocc", "Laura_S.json",
 package = "jamesdemodata", mustWork = TRUE)
js1 <- readLines(url)
js2 <- jsonlite::toJSON(jsonlite::fromJSON(url), auto_unbox = TRUE)

#' try all four inputs
m1 <- james_post(path = "upload/json", txt = fn1)
m2 <- james_post(path = "upload/json", txt = js1)
m3 <- james_post(path = "upload/json", txt = js2)
m4 <- james_post(path = "upload/json", txt = url)
}
