% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/james_post.R
\name{james_post}
\alias{james_post}
\title{Make JAMES POST request}
\usage{
james_post(path = c("version", "upload"), txt = NULL)
}
\arguments{
\item{path}{String with the path}

\item{txt}{Data set, used in \code{upload}}
}
\value{
Object of class \code{james_post}
}
\description{
Make JAMES POST request
}
\examples{
z <- james_post()
fn <- system.file("extdata", "allegrosultum", "client3.json", package = "jamesdemodata")
z <- james_post("upload", txt = fn)

url <- paste("http://localhost", "ocpu/library/jamesdemodata",
  "extdata/bds_v2.0/smocc/Laura_S.json", sep = "/")
fn <- system.file("extdata", "bds_v2.0", "smocc", "Laura_S.json",
 package = "jamesdemodata", mustWork = TRUE)
js1 <- readLines(url)
js2 <- jsonlite::toJSON(jsonlite::fromJSON(url), auto_unbox = TRUE)

z <- james_post("upload", txt = js2)
james_post()

}
