% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/james_post.R
\name{james_post}
\alias{james_post}
\title{Make JAMES POST request}
\usage{
james_post(
  host = "http://localhost",
  path = character(0),
  query = list(),
  txt = NULL,
  ...
)
}
\arguments{
\item{host}{String with the host. The default is \code{"http://localhost"}}

\item{path}{String with the path, e.g. \code{"version"}, \code{"upload"} or \code{"upload/json"}}

\item{query}{A list with query arguments, for example, \code{list(auto_unbox = TRUE)} for
\code{json} output, or \code{list(width = 7.09, height = 7.09)} for \code{svglite}.}

\item{txt}{Data set, argument used in \verb{[james::fetch_loc()]}}

\item{\dots}{Any other arguments passed to james functions via POST body}
}
\value{
Object of class \code{james_post}
}
\description{
Make JAMES POST request
}
\details{
If \code{txt} is a file, then the data are uploaded using \code{upload_file()} and
encoding \code{multipart/form}. If \code{txt} is a URL, then the data are read from the URL
as JSON string and uploaded with encoding \code{json}.
}
\examples{
url <- paste("http://localhost", "ocpu/library/jamesdemodata",
  "extdata/bds_v2.0/smocc/Laura_S.json", sep = "/")
fn <- system.file("extdata", "bds_v2.0", "smocc", "Laura_S.json",
 package = "jamesdemodata", mustWork = TRUE)
fn <- system.file("extdata", "allegrosultum", "client3.json",
 package = "jamesdemodata", mustWork = TRUE)
js <- read_json_js(txt = fn)
jo <- read_json_jo(txt = fn)

#' try all inputs
m1 <- james_post(path = "data/upload/json", txt = fn)
m2 <- james_post(path = "data/upload/json", txt = js)
m3 <- james_post(path = "data/upload/json", txt = url)

\dontrun{
# create and store A4 SVG plot
r5 <- james_post(path = "/charts/draw/svglite", txt = url,
chartcode = "NMBA", selector = "chartcode",
query = list(height = 29.7/2.54, width = 21/2.54))
tmp <- tempfile(pattern = "chart", fileext = ".svg")
writeLines(r5$content, con = tmp)
browseURL(tmp)
unlink(tmp)
}
}
