% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_chart.R
\name{request_chart}
\alias{request_chart}
\title{Request a growth chart from JAMES}
\usage{
request_chart(
  txt = NULL,
  loc = NULL,
  chartcode = NULL,
  curve_interpolation = TRUE
)
}
\arguments{
\item{txt}{A JSON string, URL or file with the data in JSON
format. The input data adhere to specification
\href{https://www.ncj.nl/themadossiers/informatisering/basisdataset/documentatie/?cat=12}{BDS JGZ 3.2.5},
and are converted to JSON according to \code{schema}.}

\item{loc}{Alternative to \code{txt}. Location where input data is uploaded
and converted to internal server format.}

\item{chartcode}{The code of the requested growth chart. If not
specified, the server will automatically plot child height for
the most recent age period.}

\item{curve_interpolation}{A logical indicating whether curve
interpolation shoud be applied.}
}
\value{
An object of class \code{\link[httr:response]{response}}
}
\description{
This function requests a growth chart from
\code{groeidiagrammen.nl}. The function can upload a JSON file
with BDS-format to the server, parse its contents into a tibble with
a person attribute, and draw the plot.
The user may skip uploading, and use a previously stored
location on the server.
}
\details{
One of \code{txt} or \code{resp} need to be specified. If both
are given, a non-NULL \code{txt} takes precedence over \code{resp}.
}
\examples{
# example with separate upload
fn <- file.path(path.package("jamesclient"), "testdata", "client3.json")
resp1 <- upload_txt(fn)
loc <- get_url(resp1, "loc")
resp2 <- request_chart(loc = loc, chartcode = "NJAA")
url <- get_url(resp2, "svglite")
# browseURL(url)

# example with integrated upload and automatic chartcode choice
resp3 <- request_chart(fn)
}
\seealso{
\code{\link[james]{draw_chart}}, \code{\link{upload_txt}}.
}
\keyword{client}
