% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_site.R
\name{request_site}
\alias{request_site}
\title{Request a site personalised chart and user nagivation}
\usage{
request_site(txt = NULL, ssd = NULL,
  host = "http://groeidiagrammen.nl")
}
\arguments{
\item{txt}{Name of file that contains valid JSON data. The
variable specification are expected to be according specification
\href{https://www.ncj.nl/themadossiers/informatisering/basisdataset/documentatie/?cat=12}{BDS
 JGZ 3.2.5}, and converted to JSON.}

\item{ssd}{Server-side data location, for example, as obtained
by \code{get_url(resp, "location")}, where \code{resp} is the
object returned by OpenCPU server.}

\item{host}{A url of the JAMES server host.}
}
\value{
An object of class \code{\link[httr:response]{response}}
}
\description{
This function requests an URL to a site that shows a personalised
growth chart plus essential user nagivation from JANES. The
function can take child data in two forms: 1) a JSON file with
BDS-formatted child data or, 2) a URL with child data on a
previously stored server location on the server.
}
\details{
One of \code{txt} or \code{resp} need to be specified. If both
are given, a non-NULL \code{txt} takes precedence over \code{resp}.
}
\examples{
# example with separate upload
library("jamesclient")
fn <- file.path(path.package("jamesclient"), "testdata", "client3.json")

# as two steps
resp <- upload_bds(fn)
ssd <- get_url(resp, "location")
site_url <- request_site(ssd = ssd)
site_url
# browseURL(site_url)

# into one step
site_url <- request_site(txt = fn)
site_url
}
\seealso{
\code{\link[james]{draw_chart}},
\code{\link{upload_bds}}.
}
\keyword{client}
