% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_bds.R
\name{upload_bds}
\alias{upload_bds}
\title{Upload and parse BDS file to ocpu server}
\usage{
upload_bds(fn = NULL, url = "http://54.93.78.215")
}
\arguments{
\item{fn}{File name of data. The variable specification are
expected to be according specification
\href{https://www.ncj.nl/themadossiers/informatisering/basisdataset/documentatie/?cat=12}{BDS
 JGZ 3.2.5}, and converted to JSON.}

\item{url}{The \code{url} of the server}
}
\value{
An object of class \code{\link[httr:response]{response}}
}
\description{
Uploads a JSON file with \code{bds} to server, parses its contents
into an object of class \linkS4class{individual}, and returns the
file keys on the server.
}
\examples{
\dontrun{
library(james.clieny)
fn <- file.path(path.package("james.client"), "testdata", "client3.json")
resp <- upload_bds(fn)
resp
}
}
\keyword{client}
